1. Moved c style functions to static methods of MacroBot class 2. Removed custom lua function 'average' (That was test purpose function)

+++ macrobot/main.cpp
@@ -18,7 +18,7 @@ int handleStdin(char ** buf)
 
 	std::cin.getline(data, BUFF_SIZE);
 
-	count = strlen(data);
+	count = (int) strlen(data);
 
 	printf("[%d]\n", count);
 	printf("[%s]\n", data);
